teste 4

+++ src/main/java/com/squad_17_api/demo/service/impl/UsuarioServiceImpl.java
@@ -1,8 +1,5 @@
 package com.squad_17_api.demo.service.impl;
 
-import com.squad_17_api.demo.exception.CursoExistenteException;
-import com.squad_17_api.demo.exception.UsuarioExistenteException;
-import com.squad_17_api.demo.model.Curso;
 import com.squad_17_api.demo.model.Usuario;
 import com.squad_17_api.demo.repository.UserRepository;
 import com.squad_17_api.demo.service.UsuarioService;
@@ -10,6 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UsuarioServiceImpl implements UsuarioService {
@@ -29,24 +27,23 @@ public class UsuarioServiceImpl implements UsuarioService {
     }
 
     @Override
+    public Optional<Usuario> buscarPorId(Integer id) {
+        return userRepository.findById(id);
+    }
+
+    @Override
     public Usuario atualizar(Usuario usuario) {
-        return userRepository.findById(usuario.getEmail())
-                .map(existingUsuario -> {
-                    existingUsuario.setEmail(usuario.getEmail());
-                    existingUsuario.setSenha(usuario.getSenha());
-                    return userRepository.save(existingUsuario);
-                })
-                .orElseGet(() -> userRepository.save(usuario));
+        return userRepository.save(usuario);
     }
 
     @Override
     public Usuario deletar(String email) {
-        return userRepository.findById(email)
-                .map(usuario -> {
-                    userRepository.delete(usuario);
-                    return usuario;
-                })
-                .orElse(null);
-    }
+        Usuario usuario = userRepository.findByEmail(email);
+        if (usuario != null) {
+            userRepository.delete(usuario);
+        }
+        return usuario;
 
+
+    }
 }
